feat: Added transition with image as shared element
Share the image in the transition between list and details activities.
Two key points: The activity has a recycler view, then the transition
name for the image has ti be set programatically since the xml layout
for the item is used by all of them (the items). So, the name is set
when in the adapter when binding the view, and the in the details fragment.

Also, since the transition is done between an activity and a fragment,
we need to make use of postponeEnterTransition and startPostponeedEnterTransition
so we make sure the transition is done when it is safe do to the animation.

Gradle module updated to lollipop (v21) to enable transitions.