fix: Refactored how/where to set transition name
In the list activity, the transition name is set using the function to return
the _ID of the movie. This is basically the same as before, just to keep
consistency.

When creating the details activity the id of the item is passed to it
encoded in the URI, this is then passed to the fragment as an argument.
Taking advantage of this, and given that the transition name is the id
of the item, the transition name is set by getting the id of the item
in the fragment's argument, rather than querying with a cursor. This
new approach guaranties that the transition name will be set. Using the
cursor was prone to errors since the cursor may be null for a number
of reasons.